REF: move Requirement into simplesat.constraints.
