+++ simplesat/constraints/__init__.py
@@ -1,2 +1,2 @@
-from .multi import MultiConstraints
 from .package_parser import PrettyPackageStringParser
+from .requirement import Requirement
