+++ simplesat/tests/test_solver.py
@@ -6,10 +6,9 @@ from okonomiyaki.versions import EnpkgVersion
 from enstaller import Repository
 from enstaller.solver import Request
 
-from simplesat.constraints import PrettyPackageStringParser
+from simplesat.constraints import PrettyPackageStringParser, Requirement
 from simplesat.dependency_solver import DependencySolver
 from simplesat.pool import Pool
-from simplesat.requirement import Requirement
 from simplesat.transaction import InstallOperation, UpdateOperation
